+++ app/Http/Controllers/Accounting/API/ProjectsController.php
@@ -64,7 +64,8 @@ class ProjectsController extends Controller
 
     private function filterQuery(Builder $query, string $filter): Builder
     {
-        return $query->where(fn (Builder $wq) => $wq
+        return $query->where(
+            fn (Builder $wq) => $wq
                 ->where('name', 'LIKE', '%'.$filter.'%')
                 ->orWhere('description', 'LIKE', '%'.$filter.'%')
         );