+++ app/Http/ViewComposers/AppVersionComposer.php
@@ -3,19 +3,18 @@
 namespace App\Http\ViewComposers;
 
 use Illuminate\View\View;
-use TJVB\LaravelGitHash\Contracts\GitHashLoader;
 
 class AppVersionComposer
 {
-    public function __construct(private GitHashLoader $gitHashLoader)
-    {
-    }
-
     /**
      * Bind data to the view.
      */
     public function compose(View $view): void
     {
-        $view->with('app_version', $this->gitHashLoader->getGitHash()->short());
+        $releaseFile = base_path('RELEASE');
+        if (is_file($releaseFile)) {
+            $version = trim(file_get_contents($releaseFile));
+            $view->with('app_version', $version);
+        }
     }
 }