+++ app/Policies/Accounting/WalletPolicy.php
@@ -20,7 +20,6 @@ class WalletPolicy
     /**
      * Determine whether the user can view any models.
      *
-     * @param  \App\Models\User  $user
      * @return mixed
      */
     public function viewAny(User $user)
@@ -31,8 +30,6 @@ class WalletPolicy
     /**
      * Determine whether the user can view the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Accounting\Wallet  $wallet
      * @return mixed
      */
     public function view(User $user, Wallet $wallet)
@@ -53,7 +50,6 @@ class WalletPolicy
     /**
      * Determine whether the user can create models.
      *
-     * @param  \App\Models\User  $user
      * @return mixed
      */
     public function create(User $user)
@@ -64,8 +60,6 @@ class WalletPolicy
     /**
      * Determine whether the user can update the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Accounting\Wallet  $wallet
      * @return mixed
      */
     public function update(User $user, Wallet $wallet)
@@ -76,8 +70,6 @@ class WalletPolicy
     /**
      * Determine whether the user can delete the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Accounting\Wallet  $wallet
      * @return mixed
      */
     public function delete(User $user, Wallet $wallet)